Fix Raw Classes and Value Classes. Make printing lambdas more explicit to also explain them
