+++ 07-presentation/src/main/scala/dagmendez/presentation/B_RawClasses.scala
@@ -10,12 +10,12 @@ object B_RawClasses extends App:
   Vector[DummyDNI](
     new DummyDNI("00000001A"), 
     new DummyDNI("ABCDEFGH0")
-  ).foreach(println)
+  ).foreach(dummyDNI => println(dummyDNI))
 
   // Basic Vanilla Scala
   case class DNI(numero: String, letra: Char):
     def value: String = s"$numero-$letra"
 
   Vector[DNI](DNI("12345678", 'A'), DNI("0", 'f'))
-    .map(_.value)
-    .foreach(println)
+    .map(dni => dni.value)
+    .foreach(dniValue => println(dniValue))
